+++ lib/LongitudeOne/Geo/String/Lexer.php
@@ -6,7 +6,7 @@
  * PHP 8.1 | 8.2 | 8.3
  *
  * Copyright LongitudeOne - Alexandre Tranchant - Derek J. Lambert.
- * Copyright 2024.
+ * Copyright 2024-2026.
  *
  */
 
@@ -72,12 +72,11 @@ class Lexer extends AbstractLexer
     protected function getType(&$value): int
     {
         if (is_numeric($value)) {
-            $value += 0;
+            $numeric = $value + 0;
 
-            if (is_int($value)) {
+            if (is_int($numeric)) {
                 return self::T_INTEGER;
             }
-            $value = (string) $value;
 
             return self::T_FLOAT;
         }