+++ lib/LongitudeOne/Geo/String/Parser.php
@@ -6,7 +6,7 @@
  * PHP 8.1 | 8.2 | 8.3
  *
  * Copyright LongitudeOne - Alexandre Tranchant - Derek J. Lambert.
- * Copyright 2024.
+ * Copyright 2024-2026.
  *
  */
 
@@ -21,6 +21,11 @@ use LongitudeOne\Geo\String\Exception\UnexpectedValueException;
  */
 class Parser
 {
+    /**
+     * Maximum length of a value embedded in an exception message.
+     */
+    private const MAX_ERROR_VALUE_LENGTH = 100;
+
     /**
      * @var string original input string
      */
@@ -181,8 +186,7 @@ class Parser
         // If degrees is a float, there will be no minutes or seconds
         if ($this->lexer->isNextToken(Lexer::T_FLOAT)) {
             // Get degree value
-            /** @var float $degrees */
-            $degrees = $this->match(Lexer::T_FLOAT);
+            $degrees = (float) $this->match(Lexer::T_FLOAT);
 
             // Degree symbol may follow degree float values
             if ($this->lexer->isNextToken(Lexer::T_DEGREE)) {
@@ -192,7 +196,6 @@ class Parser
                 $this->nextSymbol = Lexer::T_DEGREE;
             }
 
-            // Return the float value
             return $degrees;
         }
 
@@ -341,12 +344,30 @@ class Parser
         $y = $this->coordinate();
 
         // There should be no additional tokens
+        // @phpstan-ignore-next-line
         if (null !== $this->lexer->lookahead) {
             throw $this->syntaxError('end of string');
         }
 
         // Return coordinate array
         return [$x, $y];
+    }
+
+    /**
+     * Strip control characters (CR, LF, ...) and truncate a value before it is embedded in an exception message.
+     *
+     * Prevents a large or crafted input from inflating the exception message size or forging fake log lines
+     * when that message is logged as-is.
+     */
+    private function sanitizeForError(string $value): string
+    {
+        $value = (string) preg_replace('/[\x00-\x1F\x7F]/', ' ', $value);
+
+        if (strlen($value) > self::MAX_ERROR_VALUE_LENGTH) {
+            return substr($value, 0, self::MAX_ERROR_VALUE_LENGTH).'...';
+        }
+
+        return $value;
     }
 
     /**
@@ -455,14 +476,14 @@ class Parser
     {
         $expected = sprintf('Expected %s, got', $expected);
         $token = $this->lexer->lookahead;
-        $found = null === $token ? 'end of string.' : sprintf('"%s"', $token->value);
+        $found = null === $token ? 'end of string.' : sprintf('"%s"', $this->sanitizeForError((string) $token->value));
 
         $message = sprintf(
             '[Syntax Error] line 0, col %d: Error: %s %s in value "%s"',
             $token->position ?? -1,
             $expected,
             $found,
-            $this->input
+            $this->sanitizeForError($this->input)
         );
 
         return new UnexpectedValueException($message);